+++ src/error-type.ts
@@ -1,0 +1,5 @@
+export enum ErrorType {
+    ConflictingValuesForId,
+    NoAttributeFound,
+    NoInnerHTMLFound
+}