+++ src/walker.ts
@@ -1,4 +1,6 @@
 import { HTMLElement, Node } from 'node-html-parser';
+import { WarningType } from './warning-type';
+import { ErrorType } from './error-type';
 interface WalkerBaseResult {
     id: string;
     occassion: number;
@@ -9,11 +11,20 @@ interface WalkerSuccessResult extends WalkerBaseResult {
     state: 'success';
 }
 interface WalkerFailResult extends WalkerBaseResult {
-    state: 'warning' | 'error';
-    error: string;
+    message: string;
     value?: string;
 }
-export type WalkerResult = WalkerSuccessResult | WalkerFailResult;
+interface WalkerWarningResult extends WalkerFailResult {
+    state: 'warning';
+    warning: WarningType;
+}
+interface WalkerErrorResult extends WalkerFailResult {
+    state: 'error';
+    error: ErrorType;
+}
+
+
+export type WalkerResult = WalkerSuccessResult | WalkerWarningResult | WalkerErrorResult;
 
 const i18nMatchPattern = /i18n(?:-([^=]+))?/;
 
@@ -48,9 +59,9 @@ export class Walker {
                 }
                 if (!value || value.trim().length === 0) {
                     if (attribute) {
-                        result.push({ id, occassion, state: 'error', error: 'The i18n attribute was registered on a tag as attribute tag. However the matching attribute was not found!', attribute });
+                        result.push({ id, occassion, state: 'error', error: ErrorType.NoAttributeFound, message: 'The i18n attribute was registered on a tag as attribute tag. However the matching attribute was not found!', attribute });
                     } else {
-                        result.push({ id, occassion, state: 'error', error: 'The i18n attribute was registered on an element without any inner HTML', attribute });
+                        result.push({ id, occassion, state: 'error', error: ErrorType.NoInnerHTMLFound, message: 'The i18n attribute was registered on an element without any inner HTML', attribute });
                     }
                 } else {
                     result.push({ id: element.attributes[key], occassion, value: value.trim(), state: 'success', attribute });