+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## [0.2.2] - 2019-06-21
+### Added
+- Global notifications showing amount of warnings and errors in notifications when i18n changes are made
+- Shift-clicking links copies the i18n id to the clipboard
+
+
 ## [0.2.1] - 2019-06-20
 ### Bugfixes
 - Fixes tags without value to be shown and clickable in error and warning tables